feat: use watermark file and use name as uuid in document

+++ dossierfacile-api-watermark/src/main/java/fr/dossierfacile/api/pdf/service/DocumentServiceImpl.java
@@ -128,7 +128,9 @@ public class DocumentServiceImpl implements DocumentService {
         StorageFile pdfFile = document.getPdfFile();
         document.setPdfFile(null);
         watermarkDocumentRepository.save(document);
-        storageFileRepository.delete(pdfFile);
+        if (pdfFile != null) {
+            storageFileRepository.delete(pdfFile);
+        }
     }
 
     @Override
